replace erroneous comments on nodes vs randomness
The first paragraph under Discussion appears to make the common error conflating randomness with uniqueness, which entirely misunderstands probability. Including node ID enforces a uniqueness guarantee for a serial generator thread when combined with the timestamp and sub-interval incrementing. It's actually the random bits that aren't strictly needed anymore at that point.